initial code for reep.io 2.0
This is basically a reboot of reep.io

My vision for it is to write the entire application in typescript, cover everything by unit tests and improve some of the core features.

Also a big issue with current reep.io is transfer speed which is limited to around 2Mb/s for some reason. This needs to be fixed obviously. I think this is a issue with the old codebase.